--- app/src/main/java/com/dawinder/btnjc/nav/Item.kt
@@ -1,5 +1,0 @@
-package com.dawinder.btnjc.nav
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-open class Item(val path: String, val title: String, val icon: ImageVector)
