--- app/src/main/java/com/dawinder/btnjc/nav/NavPath.kt
@@ -1,5 +1,0 @@
-package com.dawinder.btnjc.nav
-
-enum class NavPath {
-    HOME, HISTORY, REMINDER, ACHIEVEMENTS, SETTINGS
-}
