--- app/src/main/java/com/dawinder/btnjc/nav/NavTitle.kt
@@ -1,9 +1,0 @@
-package com.dawinder.btnjc.nav
-
-object NavTitle {
-    const val HOME = "Home"
-    const val HISTORY = "History"
-    const val REMINDER = "Reminder"
-    const val ACHIEVEMENTS = "Goals"
-    const val SETTINGS = "Settings"
-}
