--- app/src/main/java/com/dawinder/btnjc/ui/HydroFishUIState.kt
@@ -1,7 +1,0 @@
-package com.dawinder.btnjc.ui
-
-data class HydroFishUIState(
-    val dailyWaterConsumedML: Float = 0f,
-    val curDailyMaxWaterConsumedML: Float = 1000f
-)
-
