--- app/src/main/java/com/dawinder/btnjc/ui/activities/MainActivity.kt
@@ -1,94 +1,0 @@
-package com.dawinder.btnjc.ui.activities
-
-import android.Manifest
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.os.Build
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.compose.rememberNavController
-import com.dawinder.btnjc.R
-import com.dawinder.btnjc.ui.composables.MainScreen
-import com.dawinder.btnjc.ui.theme.BottomTabNavigationJetpackComposeTheme
-
-class MainActivity : ComponentActivity() {
-    /**
-     * Called when the activity is starting. This is where most initialization should go:
-     * calling `setContentView`, instantiating UI components, and binding data.
-     *
-     * @param savedInstanceState The previously saved instance state, if any.
-     */
-
-    private lateinit var requestNotificationPermissionLauncher: ActivityResultLauncher<String>
-
-    private val showPermissionExplanationDialog = mutableStateOf(false)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        requestNotificationPermissionLauncher =
-            registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted: Boolean ->
-                if (isGranted) {
-                    // Permission is granted
-                } else {
-                    // Permission is denied
-                    if (!shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS)) {
-                        showPermissionExplanationDialog.value = true
-                    }
-                }
-            }
-
-        // Request post_notification permission
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            requestNotificationPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-        }
-
-        createNotificationChannel()
-
-        setContent {
-            BottomTabNavigationJetpackComposeTheme {
-                val navController = rememberNavController()
-                Surface(modifier = Modifier.fillMaxSize()) {
-                    MainScreen(navController = navController,showPermissionExplanationDialog.value) {
-                        showPermissionExplanationDialog.value = false
-                    }
-                }
-            }
-        }
-
-    }
-    private fun createNotificationChannel() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val name = getString(R.string.channel_name)
-            val descriptionText = getString(R.string.channel_description)
-            val importance = NotificationManager.IMPORTANCE_DEFAULT
-            val channel = NotificationChannel("alarm_channel_id", name, importance).apply {
-                description = descriptionText
-            }
-            val notificationManager: NotificationManager =
-                getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    val navController = rememberNavController()
-    BottomTabNavigationJetpackComposeTheme {
-        Surface(modifier = Modifier.fillMaxSize()) {
-            MainScreen(navController = navController,false, {} )
-        }
-    }
-}
