--- app/src/main/java/com/dawinder/btnjc/ui/composables/BottomNavigationBar.kt
@@ -1,45 +1,0 @@
-package com.dawinder.btnjc.ui.composables
-
-import androidx.compose.material3.Icon
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.navigation.NavHostController
-import com.dawinder.btnjc.nav.NavItem
-
-/**
- * Composable function that represents the bottom navigation bar of the application.
- *
- * @param navController The navigation controller used for handling navigation between screens.
- */
-@Composable
-fun BottomNavigationBar(navController: NavHostController) {
-    val navItems = listOf(NavItem.Home, NavItem.History, NavItem.Reminder, NavItem.Achievements, NavItem.Settings)
-    var selectedItem by rememberSaveable { mutableStateOf(0) }
-
-    NavigationBar {
-        navItems.forEachIndexed { index, item ->
-            NavigationBarItem(
-                alwaysShowLabel = true,
-                icon = { Icon(item.icon, contentDescription = item.title) },
-                label = { Text(item.title) },
-                selected = selectedItem == index,
-                onClick = {
-                    selectedItem = index
-                    navController.navigate(item.path) {
-                        navController.graph.startDestinationRoute?.let { route ->
-                            popUpTo(route) { saveState = true }
-                        }
-                        launchSingleTop = true
-                        restoreState = true
-                    }
-                }
-            )
-        }
-    }
-}
