--- app/src/main/java/com/dawinder/btnjc/ui/composables/tabs/SettingsScreen.kt
@@ -1,178 +1,0 @@
-package com.dawinder.btnjc.ui.composables.tabs
-
-import android.Manifest
-import android.app.AlarmManager
-import android.content.Context
-import android.content.Intent
-import android.net.Uri
-import android.os.Build
-import android.provider.Settings
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.selection.toggleable
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.unit.dp
-
-/**
- * Composable function that represents the profile screen of the application.
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun SettingsScreen() {
-    var item1EnabledState by remember { mutableStateOf(false) }
-    var item1SubItem1EnabledState by remember { mutableStateOf(false) }
-    var item1SubItem2EnabledState by remember { mutableStateOf(false) }
-    var showPermissionExplanationDialog by remember { mutableStateOf(false) }
-
-    val context = LocalContext.current
-
-    val requestNotificationPermissionLauncher = rememberLauncherForActivityResult(
-        ActivityResultContracts.RequestPermission()
-    ) { isGranted: Boolean ->
-        if (isGranted) {
-            // permission is granted, enable the switch(currently off, cuz there is other permission check)
-            //item1EnabledState = true
-        } else {
-            // permission is denied, disable the switch
-            item1EnabledState = false
-            showPermissionExplanationDialog = true
-        }
-    }
-
-    Scaffold(
-        topBar = {
-            TopAppBar(title =  { Text(text = "Settings") })
-        }
-    ) { paddingValues ->
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(paddingValues)
-        ) {
-            SettingSwitchItem(
-                title = "Notification",
-                description = "Turn on to receive notifications",
-                checked = item1EnabledState,
-                onCheckedChange = {newState ->
-                    if (newState) {
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                            // check for post_permission
-                            requestNotificationPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-                        }
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-                            val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-                            if (!alarmManager.canScheduleExactAlarms()) {
-                                item1EnabledState = false
-                                val intent = Intent(Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM)
-                                context.startActivity(intent)
-                            } else {
-                                item1EnabledState = true
-                            }
-                        } else {
-                            item1EnabledState = true
-                        }
-                    } else {
-                        item1SubItem1EnabledState = false
-                        item1SubItem2EnabledState = false
-                        item1EnabledState = false
-                    }
-                }
-            )
-            if (showPermissionExplanationDialog) {
-                PermissionExplanationDialog(onDismiss = { showPermissionExplanationDialog = false })
-            }
-        }
-    }
-}
-
-@Composable
-fun PermissionExplanationDialog(onDismiss: () -> Unit) {
-    val context = LocalContext.current
-    AlertDialog(
-        onDismissRequest = { onDismiss() },
-        title = { Text("Permission Needed") },
-        text = { Text("This app requires notification permission to function properly.") },
-        confirmButton = {
-            TextButton(onClick = {
-                val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
-                    data = Uri.fromParts("package", context.packageName, null)
-                    flags = Intent.FLAG_ACTIVITY_NEW_TASK
-                }
-                context.startActivity(intent)
-            }) {
-                Text("Open Settings")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Dismiss")
-            }
-        }
-    )
-}
-
-@Composable
-private fun SettingSwitchItem(
-    modifier: Modifier = Modifier,
-    checked: Boolean,
-    onCheckedChange: (Boolean) -> Unit,
-    title: String,
-    description: String,
-    enabled: Boolean = true,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 8.dp)
-            .toggleable(
-                value = checked,
-                enabled = enabled,
-                role = Role.Switch,
-                onValueChange = onCheckedChange
-            ),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(
-            modifier = Modifier.weight(1.0f),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-
-            Text(
-                text = title,
-                style = MaterialTheme.typography.bodyLarge,
-                maxLines = 1,
-            )
-            Text(
-                text = description,
-                style = MaterialTheme.typography.bodyMedium,
-            )
-        }
-
-        Switch(
-            checked = checked,
-            onCheckedChange = null,
-            enabled = enabled
-        )
-    }
-}
