+++ app/src/main/java/com/hydrofish/app/nav/Item.kt
@@ -1,0 +1,5 @@
+package com.hydrofish.app.nav
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+open class Item(val path: String, val title: String, val icon: ImageVector)
