+++ app/src/main/java/com/hydrofish/app/nav/NavPath.kt
@@ -1,0 +1,5 @@
+package com.hydrofish.app.nav
+
+enum class NavPath {
+    HOME, HISTORY, REMINDER, ACHIEVEMENTS, SETTINGS
+}
