+++ app/src/main/java/com/hydrofish/app/nav/NavTitle.kt
@@ -1,0 +1,9 @@
+package com.hydrofish.app.nav
+
+object NavTitle {
+    const val HOME = "Home"
+    const val HISTORY = "History"
+    const val REMINDER = "Reminder"
+    const val ACHIEVEMENTS = "Goals"
+    const val SETTINGS = "Settings"
+}
