+++ app/src/main/java/com/hydrofish/app/ui/HydroFishUIState.kt
@@ -1,0 +1,7 @@
+package com.hydrofish.app.ui
+
+data class HydroFishUIState(
+    val dailyWaterConsumedML: Float = 0f,
+    val curDailyMaxWaterConsumedML: Float = 1000f
+)
+
