+++ app/src/main/java/com/hydrofish/app/ui/theme/Color.kt
@@ -1,0 +1,62 @@
+package com.hydrofish.app.ui.theme
+import androidx.compose.ui.graphics.Color
+
+val md_theme_light_primary = Color(0xFF326B00)
+val md_theme_light_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFA1F964)
+val md_theme_light_onPrimaryContainer = Color(0xFF0B2000)
+val md_theme_light_secondary = Color(0xFF006879)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFAAEDFF)
+val md_theme_light_onSecondaryContainer = Color(0xFF001F26)
+val md_theme_light_tertiary = Color(0xFF386665)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFBBECEA)
+val md_theme_light_onTertiaryContainer = Color(0xFF002020)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF030865)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF030865)
+val md_theme_light_surfaceVariant = Color(0xFFE0E4D6)
+val md_theme_light_onSurfaceVariant = Color(0xFF44483E)
+val md_theme_light_outline = Color(0xFF74796D)
+val md_theme_light_inverseOnSurface = Color(0xFFF1EFFF)
+val md_theme_light_inverseSurface = Color(0xFF1E2578)
+val md_theme_light_inversePrimary = Color(0xFF87DC4A)
+val md_theme_light_surfaceTint = Color(0xFF326B00)
+val md_theme_light_outlineVariant = Color(0xFFC4C8BB)
+val md_theme_light_scrim = Color(0xFF000000)
+
+val md_theme_dark_primary = Color(0xFF87DC4A)
+val md_theme_dark_onPrimary = Color(0xFF173800)
+val md_theme_dark_primaryContainer = Color(0xFF245100)
+val md_theme_dark_onPrimaryContainer = Color(0xFFA1F964)
+val md_theme_dark_secondary = Color(0xFF55D6F3)
+val md_theme_dark_onSecondary = Color(0xFF003640)
+val md_theme_dark_secondaryContainer = Color(0xFF004E5C)
+val md_theme_dark_onSecondaryContainer = Color(0xFFAAEDFF)
+val md_theme_dark_tertiary = Color(0xFFA0CFCE)
+val md_theme_dark_onTertiary = Color(0xFF003737)
+val md_theme_dark_tertiaryContainer = Color(0xFF1E4E4D)
+val md_theme_dark_onTertiaryContainer = Color(0xFFBBECEA)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF030865)
+val md_theme_dark_onBackground = Color(0xFFE0E0FF)
+val md_theme_dark_surface = Color(0xFF030865)
+val md_theme_dark_onSurface = Color(0xFFE0E0FF)
+val md_theme_dark_surfaceVariant = Color(0xFF44483E)
+val md_theme_dark_onSurfaceVariant = Color(0xFFC4C8BB)
+val md_theme_dark_outline = Color(0xFF8E9286)
+val md_theme_dark_inverseOnSurface = Color(0xFF030865)
+val md_theme_dark_inverseSurface = Color(0xFFE0E0FF)
+val md_theme_dark_inversePrimary = Color(0xFF326B00)
+val md_theme_dark_surfaceTint = Color(0xFF87DC4A)
+val md_theme_dark_outlineVariant = Color(0xFF44483E)
+val md_theme_dark_scrim = Color(0xFF000000)
